feat: add FeatureGrid component to showcase technology stack with buttons
